backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.7.3
bcrypt==4.1.3
pyjwt==2.8.0
python-dotenv==1.0.1
gunicorn==22.0.0
